Split SES Lambda into two lambdas to allow each resource to be correctly accessed
