+++ judgment-reader-contact-lambda/requirements.txt
@@ -1,0 +1,4 @@
+boto3
+python-dotenv
+pytest
+pytest-cov
