+++ judgment-reader-report-lambda/requirements.txt
@@ -1,5 +1,4 @@
 psycopg2-binary
-boto3
 python-dotenv
 pytest
 pytest-cov
