--- ses_lambda/Dockerfile
@@ -1,11 +1,0 @@
-FROM public.ecr.aws/lambda/python:latest
-
-WORKDIR ${LAMBDA_TASK_ROOT}
-
-COPY requirements.txt .
-
-RUN pip install -r requirements.txt
-
-COPY ses_lambda.py .
-
-CMD ["ses_lambda.handler"]
